requirement.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
albumentations==0.5.2
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==0.4.0
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
dnf==0.0.1
docopt==0.6.2
dominate==2.4.0
entrypoints==0.3
filelock==3.0.12
fsspec==2021.4.0
future==0.18.2
fvcore==0.1.2.post20201111
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.29.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
idna==2.10
idna-ssl==1.1.0
imgaug==0.4.0
importlib-metadata==4.4.0
importlib-resources==5.1.4
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.0
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Markdown==3.3.4
MarkupSafe==2.0.1
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.2
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.0
olefile==0.44
omegaconf==2.0.6
opencv-python==4.4.0.46
opencv-python-headless==4.5.1.48
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
portalocker==2.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.1
ptyprocess==0.7.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.2
pydot==1.4.1
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pythondialog==3.5.1
pytorch-lightning
pytz==2021.1
PyYAML
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.19.2
scikit-learn==0.23.2
scipy<=1.4.1
Send2Trash==1.5.0
Shapely==1.7.1
six==1.16.0
sklearn==0.0
smmap==4.0.0
streamlit==0.82.0
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.10.0
test-tube==0.7.5
testpath==0.5.0
threadpoolctl
tifffile==2020.9.3
toml==0.10.2
toolz==0.11.1
torchfile
torchmetrics
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
typing==3.6.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
validators==0.18.2
virtualenv==20.4.7
visdom==0.1.8
watchdog==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yacs==0.1.8
yarl==1.6.3
zipp==3.4.1
argparse
kornia==0.6.8
